docker/gepa/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
dspy-ai==2.5.39
anthropic==0.45.2
httpx==0.28.1
